refactor: product view directory
